Fixed Build Path, Actually Finished Classes
