+++ it327project/src/main/java/Classes/Class.java
@@ -1,3 +1,5 @@
+package Classes;
+
 import java.time.*;
 import java.util.*;
 
@@ -9,7 +11,7 @@ public abstract class Class{
     private int creditHours;
     private String classType;
 
-    public abstract ArrayList<DayOfWeek> onDays();
+    public abstract ArrayList<DayOfWeek> getOnDays();
     public abstract String getIdentifier();
     public abstract int getStartTime();
     public abstract int getEndTime();
