+++ it327project/src/main/java/Schedule/Event.java
@@ -1,3 +1,5 @@
+package Schedule;
+
 import java.util.ArrayList;
 import java.time.DayOfWeek;
