+++ it327project/src/main/java/Schedule/Schedule.java
@@ -1,3 +1,5 @@
+package Schedule;
+
 import java.util.ArrayList;
 
 public class Schedule{
