+++ it327project/src/main/java/Schedule/ScheduleMaker.java
@@ -1,3 +1,5 @@
+package Schedule;
+
 import java.time.*;
 import java.util.ArrayList;
  public class scheduleMaker
